Move custom elements to their own folder
