+++ src/peerConnection.ts
@@ -13,13 +13,13 @@ import { updateBandwidthRestriction } from './util/sdpInject'
 import { getIceServers } from './iceServerData'
 import { onResolutionChange } from './selectDevice'
 import { updateAllParameters, updateParameters, updateResolution } from './senderParameters'
-import { localVideo } from './floating-video'
+import { localVideo } from './custom-elements/floating-video'
 import { closeShareDialog, isShareDialogOpen } from './shareDialog'
 import { stream } from './peerConnectionsManager'
 import { peerConnections } from './peerConnectionsManager'
 import { getActivePeerConnections } from './peerConnectionsManager'
 import { stateIndicator } from './peerConnectionsManager'
-import { CustomVideo } from './custom-video'
+import { CustomVideo } from './custom-elements/custom-video'
 import { addVideo, hideVideo, showVideo } from './remoteVideoManager'
 
 type PeerType = 'offer' | 'answer'
